+++ Tools/QuickEd/Classes/StringUtils.cpp
@@ -37,8 +37,8 @@ QString TruncateFileExtension(const QString& fileName, const QString& extension)
     // Just wrap around the particular DAVA engine functions.
     
     String truncatedName = fileName.toStdString();
-    
-    int truncatedStringLen = truncatedName.length() - extension.length();
+
+    size_t truncatedStringLen = truncatedName.length() - extension.length();
     bool endsWithExtension = false;
     if (fileName.length() >= extension.length())
     {
